requirements.txt
google-cloud-bigquery>=3.0,<4.0
pandas>=2.0,<3.0
numpy>=1.24,<2.0
scipy>=1.11,<2.0
scikit-learn>=1.3,<2.0
streamlit>=1.30,<2.0
plotly>=5.18,<6.0
db-dtypes>=1.1,<2.0
yfinance>=0.2.36
